requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
twilio>=8.15.0
python-dotenv>=1.1.0
requests>=2.31.0
SpeechRecognition>=3.9.0
flask-cors>=4.0.0
gunicorn>=21.2.0